+++ src/main/java/se/melhed/heroes/Warrior.java
@@ -16,14 +16,14 @@ public class Warrior extends Hero{
     }
 
     @Override
-    public int damage() {
+    public double damage() {
         double damageMultiplier = 1 + ((double)totalAttributes().getStrength() / 100);
 
         if(getEquipment().get(Slot.WEAPON) instanceof Weapon weapon) {
-            return (int) (weapon.getWeaponDamage()*damageMultiplier);
+            return weapon.getWeaponDamage()*damageMultiplier;
         }
 
-        return (int) (1 * damageMultiplier);
+        return 1 * damageMultiplier;
     }
 
     @Override
